FT basic struct
