--- evm-fungible-token/fungible-token/README.md
@@ -1,1 +1,0 @@
-# TBD
