--- evm-fungible-token/fungible-token/build.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-set -e
-cd "`dirname $0`"
-RUSTFLAGS='-C link-arg=-s' cargo build --target wasm32-unknown-unknown --release
-cp target/wasm32-unknown-unknown/release/fungible_token.wasm ./res/
