+++ evm-fungible-token/src/fungible_token/Cargo.toml
