+++ evm-fungible-token/src/fungible_token/storage_manager.rs
